improved memberships dashboard
