+++ components/memberships/membership_tiers.tsx
@@ -259,6 +259,7 @@ const MembershipTiers: React.FC<MembershipTiersProps> = ({
             </div>
           </div>
         )}
+
         <div className="mt-16 flex justify-center">
           <RadioGroup
             value={frequency}
@@ -282,7 +283,31 @@ const MembershipTiers: React.FC<MembershipTiersProps> = ({
             ))}
           </RadioGroup>
         </div>
-        <div className="isolate mx-8 mt-16 flex max-w-md flex-wrap justify-center justify-items-center gap-x-8 gap-y-8 sm:mt-20 lg:max-w-none">
+
+        <div className="isolate mx-8 mt-16 flex flex-wrap justify-center gap-x-8 gap-y-8 sm:mt-20 lg:max-w-none">
+          {memberships.length > 0 ? (
+            memberships.map((membership, index) => (
+              <MembershipCard
+                key={membership.membershipid}
+                membership={membership}
+                index={index + 1}
+                totalMemberships={memberships.length}
+                userid={userid}
+                isAuthenticated={isAuthenticated}
+                userMemberships={userMemberships}
+                handleBuyPlan={handleBuyPlan}
+                handleEditMembership={onEdit}
+                handleDeleteMembership={onDelete}
+                frequency={frequency}
+                editable={editable}
+                isCurrentPlan={currentMembershipId === membership.membershipid}
+              />
+            ))
+          ) : (
+            <p className="w-full text-center text-white">
+              No memberships available. Create one to get started!
+            </p>
+          )}
           {editable && (
             <div className="mr-16 w-full sm:w-64">
               <PlusCircleIcon
@@ -293,26 +318,9 @@ const MembershipTiers: React.FC<MembershipTiersProps> = ({
                 onClick={onCreateClick}
                 strokeWidth={2}
                 stroke="currentColor"
-              ></PlusCircleIcon>
+              />
             </div>
           )}
-          {memberships.map((membership, index) => (
-            <MembershipCard
-              key={membership.membershipid}
-              membership={membership}
-              index={index + 1}
-              totalMemberships={memberships.length + 1}
-              userid={userid}
-              isAuthenticated={isAuthenticated}
-              userMemberships={userMemberships}
-              handleBuyPlan={handleBuyPlan}
-              handleEditMembership={onEdit}
-              handleDeleteMembership={onDelete}
-              frequency={frequency}
-              editable={editable}
-              isCurrentPlan={currentMembershipId === membership.membershipid}
-            />
-          ))}
         </div>
       </div>
       <ToastContainer position="bottom-right" autoClose={3000} />
